Fixing validation schema loading when using CacheResourceLoader
A test with a complex api (rp-590) that contains multiple schemas and
examples was added.
CacheResourceLoader was improved, now it was modified in order to call
callback it receives in the fetch resource method.

The problem with CacheResourceLoader first version was that for
validation of json's examples with their respective schemas it is needed
an absolute URI that was set within a callback in
RootRamlFileResourceLoader, but when it was retrieved from
CacheResourceLoader that callback was not called so then validation was
done with the last correctly set schema.