Add Automatic-Module-Name for Java Modules
"2" is not a allowed module name and the default way of turning
jar-names into Java modules translated into "raml.parser.2" which is not
a valid module name, thus preventing this library to be used in a Java
module project. This introduced a MANIFEST.MF entry recommended by the
Java system to explicitly name the module (removing "2") and this
entry should have no affect on exiting non-module projects.